move aiddesfire to mifare folder and refactoring
